csv-trigger-app/gcpfunction/requirements.txt
functions-framework
firebase-admin
google-cloud-storage
python-decouple
